Badminton_Venue_Booking/venue_booking/requirements.txt
Django==5.1
djangorestframework==3.15.2
